refactor(NextButtonOffices): changed to camel case, changed div to h5, transferred to offices folder
