+++ src/app/NextButtonForOffices/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import NextButtonForOffices from '@/components/ui/Offices/OfficesFloatingCard/NextButtonForOffices';
+import NextButtonForOffices from '@/components/ui/Offices/NextButtonForOffices';
 const page = () => {
   return <NextButtonForOffices office="OVP" />;
 };
