--- src/app/modal-component/page.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-
-import Image from 'public/assets/images/HolisticHealth.png';
-
-import ModalOverlay from '@/components/ui/ModalOverlayComponent/ModalComponent';
-
-const page = () => {
-  return (
-    <ModalOverlay
-      imageUrl={Image}
-      hashtag="#testTagline"
-      title="Holistic Health"
-      description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
-      link="/"
-    />
-  );
-};
-
-export default page;
